Support multiplayer players without roster entries
It is now possible to assign an anonymous pilot for multiplayer (not
solo) games. It works like a normal human pilot except that nothing is
saved in the roster for this player.

Also fixes a bug where the up/down arrows in the Assign Players menu
allowed switching to an empty player name (which was removed when
exiting the menu).