Basic network game support
Support for playing over the network (LAN or Internet, but the server
needs to be able to receive incoming TCP connections). It is
controlled via a new Network game menu and new parts of the Assign
players menu. The network game works, but is a bit unpolished, and
there is no documentation yet.

The network game uses the zlib compression library, which is now
needed to compile Triplane.

This commit adds new network game-related fields to the end of the
Triplane configuration file triplane.cfg. Old configuration files are
read and converted automatically.