Check for ping timeouts in network games
The server now send ping messages about every two seconds, and
disconnects clients if they have not replied in about 20 seconds.

This is done to get a more accurate view of which players are
connected. More importantly, if a client or network connection gets
stuck for some reason (without the TCP connection closing cleanly),
the ping timeout allows the player to reconnect using the same player
name (and thus continue controlling a plane) after waiting for the 20
seconds.